学习

+++ app/src/main/java/com/business/travel/app/dal/entity/User.java
@@ -9,16 +9,15 @@ import androidx.room.PrimaryKey;
 @Entity
 public class User {
 	@PrimaryKey(autoGenerate = true)
-	public int id;
-
+	public long id;
 	public String firstName;
 	public String lastName;
 
-	public int getId() {
+	public long getId() {
 		return id;
 	}
 
-	public void setId(int id) {
+	public void setId(long id) {
 		this.id = id;
 	}
